Fix for gauche 0.9.9 (#12)
* Import hash-table-for-each from (gauche base)

The (gauche hashutil) library isn't meant to be used by the user program;
it is autoloaded when hash-table-for-each is accessed.

In Gauche 0.9.9, hash-table-for-each will be moved to the core, and
this change makes gauche-swank work in both 0.9.9 and before.

* Avoid depending on overriding definition behavior

Saving older definition and then redefine it has been undefined bevior,
which just happened to work up to Gauche 0.9.8.  As of 0.9.9, the binding
will be inserted at the compile time rather than runtime, and the previous
code won't work.