[guile] $condition-links and $condition-location
This is probably not implemented the way it was supposed to be, but
seems to work. I didn't find a way of getting a backtrace out of a condition,
instead I store the backtrace as soon the condition is thrown, and access
it from other procedures.

I'm not sure what $condition-links is meant to store, probably continuations
of a condition? Instead I just stored the frame locations.