Improve guile support (#14)
* update swank version

* implement basic $binding-documentation and $set-package for guile

* basic error message

* experimenting with backtraces for Guile

* add instruction for guile to README

* try improving stack trace for guile

* [guile] improve backtrace; add object/class introspection

* [guile] implement $frame-locals-and-catch-tags

* [guile] record inspection

* [guile] $condition-links and $condition-location

This is probably not implemented the way it was supposed to be, but
seems to work. I didn't find a way of getting a backtrace out of a condition,
instead I store the backtrace as soon the condition is thrown, and access
it from other procedures.

I'm not sure what $condition-links is meant to store, probably continuations
of a condition? Instead I just stored the frame locations.

* [guile] simplify object inspectors

* [guile] handle error messages with multiple irritants

* reverse arguments of intersperse