Don't use prompt to limit backtrace.
Using prompt seemed to be a good idea to limit Guile's backtrace. Unfortunately
it would mean, that one would have to set prompt on every command  evaluating
code (like slime:load-file). A simpler approach is to narrow the stack to the
most recent stack-stack. The backtrace has more stuff than when using the REPL,
but it's IMHO not a big deal.